Update code version generation logic
